Subtotal & Total methods

+++ src/main/java/com/sutanrrier/curso/resources/OrderResource.java
@@ -25,7 +25,7 @@ public class OrderResource {
 	}
 	
 	@GetMapping(value="/{id}")
-	public ResponseEntity<Order> findByiD(@PathVariable Integer id){
+	public ResponseEntity<Order> findByiD(@PathVariable Long id){
 		Order obj = service.findById(id);
 		return ResponseEntity.ok().body(obj);
 	}
